added systemtimestamp servlet
To synchronise annotations (on WB) and their encodings (stored IA side) we 1) need timestamps associated to annotations (or features) and the corresponding encodings 2) Want these timestamps to come from a central source (wb server) 3) need timestamp after IA has detected features but before callback is sent (so encodings can first be created, timestamped and saved IA side, so we can report error if there's a problem with encoding). NB if / when we have users modifying / creating features, timestamp needs updating as well as call to IA, with timestamp, to update it's cached encodings or detections etc.